updated tests, resolved conflicts
